+++ config/CHANGELOG.md
@@ -11,6 +11,28 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 [comment]: <> (Fixed:      any bug fixes)
 [comment]: <> (Security:   in case of vulnerabilities)
 
+### 1.4.5 -
+### casper-node 1.4.5-a7f6a648d-casper-mainnet
+
+## Added
+* chainspec.toml
+  * core
+    * max_stored_value_size
+    * minimum_delegation_amount
+  * deploys
+    * block_max_approval_count
+
+## Changed 
+* chainspec.toml
+  * protocol
+    * version
+    * activation_point
+* config-example.toml
+  * network
+    * known_addresses
+  * contract_runtime
+    * max_global_state_size
+
 ## 1.4.4 - 3435
 ### casper-node 1.4.4-6962edb3d-casper-mainnet